Added AssetCommitment.Validate, redefined NonconfidentialIARP.Validate in terms of it.
Added Y to ConfidentialIARP, renamed TracingPoint to T (to match the
spec), renamed ConfidentialIARP.IssuanceProof to IssuanceZKP (to avoid
confusion with the separate IssuanceProof type).

Did away with the AssetIssuanceKeyTuple type, which was intended to
reduce the burden on callers but actually increased it. Tweaked a
couple of other types for consistency with the rest of the crypto/ca
code and with the spec, and for type safety.